+++ App/ACF.php
@@ -21,19 +21,6 @@ class ACF implements WordPressHooks {
      */
     public function addHooks() {
         add_action( 'init', [ $this, 'addOptionsPage' ] );
-        add_action( 'admin_head', [ $this, 'removePageBuilderEditor' ], 100 );
-    }
-
-    /**
-     * Removes default page editor for page builer templates
-     */
-    public function removePageBuilderEditor()
-    {
-      $template_file = basename( get_page_template() );
-
-       if($template_file === 'template-page-builder.php') { // template
-         remove_post_type_support('page', 'editor');
-       }
     }